+++ gdsfactory/tests/test_validator.py
@@ -24,7 +24,7 @@ def test_validator_pass():
 
 
 def test_validator_fail_name_too_long():
-    component = gf.Component(name="a" * 33)
+    component = gf.Component(name="a" * 200)
 
     # component_with_straight(component=component)
     with pytest.raises(pydantic.ValidationError):